Add documentation describing the functionality of the HPP plot web app and improve code formatting
