+++ src/pyhpp_plot/interactive_viewer.py
@@ -5,52 +5,6 @@ Python code to add custom actions to context menus for nodes and edges.
 Implements all features from the CORBA hpp-monitoring-plugin.
 """
 
-import http.server
-import os
-import socket
-import threading
-import time
-from pathlib import Path
-
-from .websocket_bridge import GraphWebSocketBridge, _serialize_graph
-
-
-def _find_webapp_dist() -> Path | None:
-    """Locate the built React app (dist/)."""
-
-    module_path = Path(__file__).resolve()
-    for depth in range(3, 7):
-        try:
-            prefix = module_path.parents[depth]
-        except IndexError:
-            break
-        candidate = prefix / "share" / "hpp-plot" / "webapp"
-        if (candidate / "index.html").is_file():
-            return candidate
-
-    return None
-
-
-def _serve_webapp(dist_dir: Path, host: str, port: int) -> http.server.HTTPServer:
-    """Serve static files from dist_dir using stdlib HTTP server.
-
-    Returns the HTTPServer instance (already started in a daemon thread).
-    Logs HTTP requests are suppressed.
-    """
-
-    class _QuietHandler(http.server.SimpleHTTPRequestHandler):
-        def __init__(self, *args, **kwargs):
-            super().__init__(*args, directory=str(dist_dir), **kwargs)
-
-        def log_message(self, *args):
-            pass
-
-    server = http.server.HTTPServer((host, port), _QuietHandler)
-    t = threading.Thread(target=server.serve_forever, daemon=True)
-    t.name = "ReactStaticServer"
-    t.start()
-    return server
-
 
 class InteractiveGraphViewer:
     """Wrapper for HppNativeGraphWidget with Python-based context menu actions."""
@@ -298,7 +252,7 @@ class InteractiveGraphViewer:
 
         return None
 
-    def handle_react_message(self, message):
+    def handle_web_app_message(self, message):
         """Handle websocket actions coming from the React app."""
         if not isinstance(message, dict):
             return False
@@ -346,126 +300,3 @@ class InteractiveGraphViewer:
             return False
 
         return False
-
-
-class GraphViewerThread(threading.Thread):
-    """Runs graph viewer in separate daemon"""
-
-    def __init__(
-        self,
-        graph,
-        problem,
-        config_callback,
-        ws_host="127.0.0.1",
-        ws_port=8765,
-        start_qt_viewer=False,
-        react_host="127.0.0.1",
-        react_port=5173,
-    ):
-        super().__init__(daemon=True, name="GraphViewerThread")
-        self.graph = graph
-        self.problem = problem
-        self.config_callback = config_callback
-        self.ws_host = ws_host
-        self.ws_port = ws_port
-        self.start_qt_viewer = start_qt_viewer
-        self.react_host = react_host
-        self.react_port = react_port
-        self._viewer = None
-        self._ws_bridge = None
-        self._http_server = None
-        self._lock = threading.Lock()
-        self._stop_event = threading.Event()
-
-    def run(self):
-        self._viewer = InteractiveGraphViewer(self.graph, self.problem, self._on_config_generated)
-
-        if self.start_qt_viewer:
-            self._viewer.show()
-
-        else:
-            ## start webAPP
-            self._start_web_app()
-
-
-            ## start websocket bridge
-            self._ws_bridge = GraphWebSocketBridge(host=self.ws_host,port=self.ws_port,on_message=self.handle_react_message,snapshot_provider=self._build_snapshot_payload)
-            self._ws_bridge.start()
-            if self._ws_bridge.is_running:
-                self._ws_bridge.send_viewer_snapshot(self.graph, self.problem)
-                
-            try:
-                while ( not self._stop_event.is_set()
-                        and self._ws_bridge is not None
-                        and self._ws_bridge.is_running):
-                            time.sleep(0.2)
-            finally:
-                if self._ws_bridge is not None:
-                    self._ws_bridge.stop()
-                if self._http_server is not None:
-                    self._http_server.shutdown()
-                    self._http_server = None
-      
-
-
-    def _on_config_generated(self, config, label):
-        self.config_callback(config, label)
-        self.send_config(config, label)
-
-
-    def _build_snapshot_payload(self):
-        return {
-            "type": "viewer_snapshot",
-            "graph": _serialize_graph(self.graph),
-        }
-
-
-
-    def _is_tcp_port_free(self, host, port, timeout=0.2):
-        with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
-            sock.settimeout(timeout)
-            return sock.connect_ex((host, int(port))) != 0
-
-    def _start_web_app(self):
-
-        if not self._is_tcp_port_free(self.react_host, self.react_port):
-            print(f"Port {self.react_port} is not free, skipping React app auto-start.")
-            return
-
-        dist_dir = _find_webapp_dist()
-        if dist_dir is None:
-            print("React app dist/ not found, skipping auto-start.\n")
-            return
-        self._http_server = _serve_webapp(dist_dir, self.react_host, self.react_port)
-        print(f"React app ready at http://{self.react_host}:{self.react_port}")
-
-
-
-    @property
-    def ws_bridge(self):
-        return self._ws_bridge
-
-    @property
-    def is_ws_running(self):
-        return self._ws_bridge is not None and self._ws_bridge.is_running
-
-    def send_viewer_snapshot(self, graph, problem):
-        if self._ws_bridge is None or not self._ws_bridge.is_running:
-            return
-        self._ws_bridge.send_viewer_snapshot(graph, problem) 
-
-    def send_config(self, config, label):
-        if self._ws_bridge is None or not self._ws_bridge.is_running:
-            return
-        self._ws_bridge.send_config(config, label)
-
-    def handle_react_message(self, message):
-        """Dispatch a React websocket message to the interactive viewer."""
-        if self._viewer is None:
-            return False
-        with self._lock:
-            return self._viewer.handle_react_message(message)
-
-    def stop(self):
-        """Request the thread to stop (headless mode)."""
-        self._stop_event.set()
