+++ src/web_app/src/graph/normalizeSnapshot.js
@@ -2,9 +2,8 @@ function inferWaypointNodeIds(snapshotEdges) {
   const waypointNodeIds = new Set();
   const edges = Array.isArray(snapshotEdges) ? snapshotEdges : [];
 
-  const waypointParents = edges.filter(
-    (edge) => edge && edge.nbWaypoints > 0 && edge.id != null
-  );
+  const waypointParents =
+      edges.filter((edge) => edge && edge.nbWaypoints > 0 && edge.id != null);
 
   waypointParents.forEach((parent) => {
     const parentEndpoints = new Set([
@@ -18,8 +17,10 @@ function inferWaypointNodeIds(snapshotEdges) {
       const sourceId = edge.source != null ? String(edge.source) : null;
       const targetId = edge.target != null ? String(edge.target) : null;
 
-      if (sourceId && !parentEndpoints.has(sourceId)) waypointNodeIds.add(sourceId);
-      if (targetId && !parentEndpoints.has(targetId)) waypointNodeIds.add(targetId);
+      if (sourceId && !parentEndpoints.has(sourceId))
+        waypointNodeIds.add(sourceId);
+      if (targetId && !parentEndpoints.has(targetId))
+        waypointNodeIds.add(targetId);
     });
   });
 
@@ -31,8 +32,10 @@ export function elementsFromGraphSnapshot(graphSnapshot) {
     return null;
   }
 
-  const states = Array.isArray(graphSnapshot.states) ? graphSnapshot.states : [];
-  const snapshotEdges = Array.isArray(graphSnapshot.edges) ? graphSnapshot.edges : [];
+  const states =
+      Array.isArray(graphSnapshot.states) ? graphSnapshot.states : [];
+  const snapshotEdges =
+      Array.isArray(graphSnapshot.edges) ? graphSnapshot.edges : [];
   const inferredWaypointNodeIds = inferWaypointNodeIds(snapshotEdges);
 
   const nodeIds = new Set();
@@ -43,31 +46,34 @@ export function elementsFromGraphSnapshot(graphSnapshot) {
     const isWaypoint = inferredWaypointNodeIds.has(nodeId);
     return {
       data: {
-        type: isWaypoint ? "WaypointState" : "State",
+        type: isWaypoint ? 'WaypointState' : 'State',
         id: nodeId,
         label: state.name,
         name: state.name,
         constraints: state.constraints,
         constraints_functions: state.numericalConstraints,
       },
-      classes: isWaypoint ? "waypoint" : "state",
+      classes: isWaypoint ? 'waypoint' : 'state',
     };
   });
 
   const edges = snapshotEdges.map((edge) => {
     const classNames = [];
-    if (String(edge.source) === String(edge.target)) classNames.push("self-loop");
-    if (Number(edge.weight) <= 0) classNames.push("dotted");
+    if (String(edge.source) === String(edge.target))
+      classNames.push('self-loop');
+    if (Number(edge.weight) <= 0) classNames.push('dotted');
 
     return {
-      classes: classNames.join(" "),
+      classes: classNames.join(' '),
       data: {
-        type: "Edge",
+        type: 'Edge',
         id: String(edge.id),
         source: String(edge.source),
         target: String(edge.target),
         weight: Number(edge.weight),
-        controlPointStepSize: String(edge.source).length > 10 ? String(edge.source).length * 2 : 40,
+        controlPointStepSize: String(edge.source).length > 10 ?
+            String(edge.source).length * 2 :
+            40,
         label: String(edge.id),
         name: edge.name,
         nbWaypoints: edge.nbWaypoints,
