+++ src/web_app/vite.config.js
@@ -1,5 +1,5 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
+import react from '@vitejs/plugin-react';
+import {defineConfig} from 'vite';
 
 export default defineConfig({
   plugins: [react()],
